updated analyze position Step. Still needs to be tested and debugged
is correctly subclassed from the Abtract NGSStep class

created separate SAMEntry class to parse SAM line as this is required in more than one class
created MappedRead class to store the relevant information for each mapped read for subsequent analysis

this creates more overhead, but it makes the code cleaner. It should be fine for smallRNA projects.